add support for blob, timestamp (based on INT96). Cast into smaller types (todo fix signedness and handle nil in values space)

+++ sql/backends/monet5/vaults/parquet/pqc_reader.h
@@ -18,7 +18,7 @@ pqc_export pqc_reader_t *pqc_reader( pqc_reader_t *p, pqc_file *pq, int nr_worke
 pqc_export void pqc_reader_destroy( pqc_reader_t *r);
 
 pqc_export int64_t pqc_mark_chunk( pqc_reader_t *r, int nr_workers, int wnr, uint64_t nrows);
-pqc_export int64_t pqc_read_chunk( pqc_reader_t *r, int wnr, void *d, void *vd, uint64_t nrows, int *ssize, int *dict);
+pqc_export int64_t pqc_read_chunk( pqc_reader_t *r, int wnr, void *d, void *vd, uint64_t nrows, size_t *ssize, int *dict);
 
 pqc_export char *pqc_get_error( pqc_reader_t *r);
